requirements.txt
# Core dependencies for HVAC Insulation Estimator SaaS
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# AI-powered estimation with Claude Agents SDK
anthropic>=0.40.0
python-dotenv>=1.0.0

# AI-powered PDF extraction (legacy support)
google-generativeai>=0.8.0
pillow>=10.0.0
pdf2image>=1.16.0

# PDF processing (optimized)
pdfplumber>=0.10.0
opencv-python-headless>=4.8.0
pymupdf>=1.23.0  # Fast PDF rendering and text extraction

# Data validation
pydantic>=2.0.0

# Excel file support
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Performance and async
aiofiles>=23.0.0  # Async file operations

# For future enhancements
plotly>=5.18.0  # Interactive charts
altair>=5.2.0   # Declarative visualizations
